+++ lib/mongoid/paranoia.rb
@@ -80,6 +80,7 @@ module Mongoid
     def destroyed?
       (@destroyed ||= false) || !!deleted_at
     end
+    alias :deleted? :destroyed?
 
     # Restores a previously soft-deleted document. Handles this by removing the
     # deleted_at flag.